Address Merge comments
