+++ mitreattack/navlayers/core/__init__.py
@@ -8,3 +8,4 @@ from .legenditem import LegendItem
 from .metadata import Metadata
 from .technique import Technique
 from .versions import Versions
+from .helpers import *
