+++ mitreattack/navlayers/core/layout.py
@@ -50,12 +50,12 @@ class Layout:
                 return None
             if all(x == 0 for x in scores):
                 return 0
-            modified = self.aggFunction(scores, total)
+            modified = self._aggFunction(scores, total)
             return modified
         else:
             return None
 
-    def aggFunction(self, data_block, total_techs):
+    def _aggFunction(self, data_block, total_techs):
         """
             Apply the selected aggregate function to a data_block
             :param data_block: List containing scores to compute over (generated by compute_aggregate)
