+++ mitreattack/navlayers/core/technique.py
@@ -1,11 +1,13 @@
 try:
     from ..core.exceptions import BadInput, handler, typeChecker, \
-        UNSETVALUE, UnknownTechniqueProperty, BadType, handle_object_placement
+        UNSETVALUE, UnknownTechniqueProperty, BadType
     from ..core.metadata import Metadata, MetaDiv
+    from ..core.helpers import handle_object_placement
 except ImportError:
     from core.exceptions import BadInput, handler, typeChecker, \
-        UNSETVALUE, UnknownTechniqueProperty, BadType, handle_object_placement
+        UNSETVALUE, UnknownTechniqueProperty, BadType
     from core.metadata import Metadata, MetaDiv
+    from core.helpers import handle_object_placement
 
 
 class Technique:
